chore: compact contexts + providers
