+++ src/contexts/AuthContext.tsx
@@ -1,8 +1,22 @@
-import { ReactNode, useCallback, useEffect, useState } from 'react'
-import { fetchAccount, connectLichessUrl, logoutAndFetchAccount } from 'src/api'
-import {} from 'src/components'
-import { AuthContext } from 'src/contexts'
 import { User } from 'src/types'
+import React, { ReactNode, useCallback, useEffect, useState } from 'react'
+import { fetchAccount, connectLichessUrl, logoutAndFetchAccount } from 'src/api'
+
+interface IAuthContext {
+  user: User | null
+  connectLichess: () => void
+  logout: () => Promise<void>
+}
+
+export const AuthContext = React.createContext<IAuthContext>({
+  user: null,
+  connectLichess: () => {
+    throw new Error('poorly provided AuthContext, missing connectLichess')
+  },
+  logout: () => {
+    throw new Error('poorly provided AuthContext, missing logout')
+  },
+})
 
 export const AuthContextProvider: React.FC<{ children: ReactNode }> = ({
   children,
