--- src/contexts/ModalContext.ts
@@ -1,15 +1,0 @@
-import React, { ComponentProps } from 'react'
-import { PlaySetupModal } from 'src/components'
-
-const fn = () => {
-  throw new Error('poorly provided ModalContext')
-}
-
-interface IModalContext {
-  playSetupModalProps?: ComponentProps<typeof PlaySetupModal>
-  setPlaySetupModalProps: (arg0?: ComponentProps<typeof PlaySetupModal>) => void
-}
-
-export const ModalContext = React.createContext<IModalContext>({
-  setPlaySetupModalProps: fn,
-})
